refactor(google-sheets): rename modules for better clarity and consistency
- Rename FormDataFormatter to formatForSpreadsheet

- Rename GoogleSheetsAuthProvider to getSheetsAuthConfig

- Update test files and imports accordingly
