+++ website/modules/@apostrophecms/form/lib/getSheetsAuthConfig.js
